store-in-gcs/requirements.txt
bs4
functions-framework
google-cloud-storage
google-events
